add ProductValidator and CategoryValidator

+++ src/main/java/com/lpatros/ecommerce_api/dto/product/ProductResponse.java
@@ -18,7 +18,7 @@ public class ProductResponse {
     private Long id;
     private String name;
     private String description;
-    private Long stock;
+    private Integer stock;
     private BigDecimal price;
     private List<ProductImageResponse> productImages;
     private LocalDateTime createdAt;
